backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyjwt[crypto]==2.8.0
python-jose[cryptography]==3.3.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
